+++ c8y_api/model/events.py
@@ -10,7 +10,7 @@ from typing import Generator, List, BinaryIO
 
 from c8y_api._base_api import CumulocityRestApi
 from c8y_api.model._base import CumulocityResource, SimpleObject, ComplexObject
-from c8y_api.model._parser import as_tuples as parse_as_tuples, ComplexObjectParser
+from c8y_api.model._parser import as_values as parse_as_values, ComplexObjectParser
 from c8y_api.model._util import _DateUtil
 
 
@@ -264,7 +264,7 @@ class Events(CumulocityResource):
                reverse: bool = False, limit: int = None,
                with_source_assets: bool = None, with_source_devices: bool = None,
                page_size: int = 1000, page_number: int = None,
-               as_tuples: str | tuple | list[str|tuple] = None,
+               as_values: str | tuple | list[str | tuple] = None,
                **kwargs) -> Generator[Event]:
         """Query the database for events and iterate over the results.
 
@@ -315,7 +315,7 @@ class Events(CumulocityResource):
                 parsed in one chunk). This is a performance related setting.
             page_number (int): Pull a specific page; this effectively disables
                 automatic follow-up page retrieval.
-            as_tuples: (*str|tuple):  Don't parse objects, but directly extract
+            as_values: (*str|tuple):  Don't parse objects, but directly extract
                 the values at certain JSON paths as tuples; If the path is not
                 defined in a result, None is used; Specify a tuple to define
                 a proper default value for each path.
@@ -344,8 +344,8 @@ class Events(CumulocityResource):
             base_query,
             page_number,
             limit,
-            Event.from_json if not as_tuples else
-            lambda x: parse_as_tuples(x, as_tuples))
+            Event.from_json if not as_values else
+            lambda x: parse_as_values(x, as_values))
 
     def get_all(
             self,
@@ -362,7 +362,7 @@ class Events(CumulocityResource):
             reverse: bool = False, limit: int = None,
             with_source_assets: bool = None, with_source_devices: bool = None,
             page_size: int = 1000, page_number: int = None,
-            as_tuples: str | tuple | list[str|tuple] = None,
+            as_values: str | tuple | list[str | tuple] = None,
             **kwargs) -> List[Event]:
         """Query the database for events and return the results as list.
 
@@ -388,7 +388,7 @@ class Events(CumulocityResource):
             reverse=reverse,
             with_source_devices=with_source_devices, with_source_assets=with_source_assets,
             limit=limit, page_size=page_size, page_number=page_number,
-            as_tuples=as_tuples,
+            as_values=as_values,
             **kwargs
         ))