+++ c8y_api/model/measurements.py
@@ -9,7 +9,7 @@ from typing import Type, List, Generator, Sequence
 from c8y_api._base_api import CumulocityRestApi
 
 from c8y_api.model._base import CumulocityResource, ComplexObject
-from c8y_api.model._parser import as_tuples as parse_as_tuples, ComplexObjectParser
+from c8y_api.model._parser import as_values as parse_as_values, ComplexObjectParser
 from c8y_api.model._base import _DictWrapper
 from c8y_api.model._util import _DateUtil
 
@@ -495,7 +495,7 @@ class Measurements(CumulocityResource):
             limit: int = None,
             page_size: int = 1000,
             page_number: int = None,
-            as_tuples: str | tuple | list[str|tuple] = None,
+            as_values: str | tuple | list[str | tuple] = None,
             **kwargs
     ) -> Generator[Measurement]:
         """ Query the database for measurements and iterate over the results.
@@ -539,7 +539,7 @@ class Measurements(CumulocityResource):
                 related setting.
             page_number (int): Pull a specific page; this effectively disables
                 automatic follow-up page retrieval.
-            as_tuples: (*str|tuple):  Don't parse objects, but directly extract
+            as_values: (*str|tuple):  Don't parse objects, but directly extract
                 the values at certain JSON paths as tuples; If the path is not
                 defined in a result, None is used; Specify a tuple to define
                 a proper default value for each path.
@@ -568,8 +568,8 @@ class Measurements(CumulocityResource):
             base_query,
             page_number,
             limit,
-            Measurement.from_json if not as_tuples else
-            lambda x: parse_as_tuples(x, as_tuples))
+            Measurement.from_json if not as_values else
+            lambda x: parse_as_values(x, as_values))
 
     def get_all(
             self,
@@ -589,7 +589,7 @@ class Measurements(CumulocityResource):
             limit: int = None,
             page_size: int = 1000,
             page_number: int = None,
-            as_tuples: str | tuple | list[str|tuple] = None,
+            as_values: str | tuple | list[str | tuple] = None,
             **kwargs
     ) -> List[Measurement]:
         """ Query the database for measurements and return the results
@@ -618,7 +618,7 @@ class Measurements(CumulocityResource):
             limit=limit,
             page_size=page_size,
             page_number=page_number,
-            as_tuples=as_tuples,
+            as_values=as_values,
             **kwargs))
 
     def get_last(