+++ c8y_api/model/operations.py
@@ -8,7 +8,7 @@ from typing import Type, List, Generator
 from c8y_api._base_api import CumulocityRestApi
 
 from c8y_api.model._base import CumulocityResource, ComplexObject, SimpleObject, _DictWrapper
-from c8y_api.model._parser import ComplexObjectParser, as_tuples as parse_as_tuples
+from c8y_api.model._parser import ComplexObjectParser, as_values as parse_as_values
 from c8y_api.model._util import _DateUtil
 
 
@@ -161,7 +161,7 @@ class Operations(CumulocityResource):
             min_age: timedelta = None, max_age: timedelta = None,
             reverse: bool = False, limit: int = None,
             page_size: int = 1000, page_number: int = None,
-            as_tuples: str | tuple | list[str|tuple] = None,
+            as_values: str | tuple | list[str | tuple] = None,
             **kwargs
     ) -> Generator[Operation]:
         """ Query the database for operations and iterate over the results.
@@ -200,7 +200,7 @@ class Operations(CumulocityResource):
                 related setting.
             page_number (int): Pull a specific page; this effectively disables
                 automatic follow-up page retrieval.
-            as_tuples: (*str|tuple):  Don't parse objects, but directly extract
+            as_values: (*str|tuple):  Don't parse objects, but directly extract
                 the values at certain JSON paths as tuples; If the path is not
                 defined in a result, None is used; Specify a tuple to define
                 a proper default value for each path.
@@ -220,8 +220,8 @@ class Operations(CumulocityResource):
             base_query,
             page_number,
             limit,
-            Operation.from_json if not as_tuples else
-            lambda x: parse_as_tuples(x, as_tuples))
+            Operation.from_json if not as_values else
+            lambda x: parse_as_values(x, as_values))
 
     def get_all(
             self,
@@ -232,7 +232,7 @@ class Operations(CumulocityResource):
             min_age: timedelta = None, max_age: timedelta = None,
             reverse: bool = False, limit: int = None,
             page_size: int = 1000, page_number: int = None,
-            as_tuples: str | tuple | list[str|tuple] = None,
+            as_values: str | tuple | list[str | tuple] = None,
             **kwargs
     ) -> List[Operation]:
         """ Query the database for operations and return the results
@@ -248,7 +248,7 @@ class Operations(CumulocityResource):
             expression=expression,
             agent_id=agent_id, device_id=device_id, status=status, bulk_id=bulk_id,
             fragment=fragment, before=before, after=after, min_age=min_age, max_age=max_age,
-            reverse=reverse, limit=limit, page_size=page_size, page_number=page_number, as_tuples=as_tuples,
+            reverse=reverse, limit=limit, page_size=page_size, page_number=page_number, as_values=as_values,
             **kwargs
         ))