+++ tests/model/test_operations.py
@@ -62,8 +62,8 @@ def test_select(fun, params, expected, not_expected):
         assert ne not in resource
 
 
-def test_select_as_tuples():
-    """Verify that select as tuples works as expected."""
+def test_select_as_values():
+    """Verify that select as values works as expected."""
     jsons = [
         {'type': 'type1',
          'creationTime': 'time1',
@@ -78,7 +78,7 @@ def test_select_as_tuples():
 
     api = Operations(c8y=Mock())
     api.c8y.get = Mock(side_effect=[{'operations': jsons}, {'operations': []}])
-    result = api.get_all(as_tuples=[
+    result = api.get_all(as_values=[
         'type', 'creation_time', 'c8y_Command.text', 'c8y_Command.key2', 'description'])
     assert result == [
         ('type1', 'time1', 'value1', 'value2', None),
@@ -86,7 +86,7 @@ def test_select_as_tuples():
     ]
 
     api.c8y.get = Mock(side_effect=[{'operations': jsons}, {'operations': []}])
-    result = api.get_all(as_tuples=[
+    result = api.get_all(as_values=[
         'type', 'creation_time', 'c8y_Command.text', 'c8y_Command.key2', ('description', '-')])
     assert result == [
         ('type1', 'time1', 'value1', 'value2', '-'),
@@ -94,10 +94,10 @@ def test_select_as_tuples():
     ]
 
     api.c8y.get = Mock(side_effect=[{'operations': jsons}, {'operations': []}])
-    result = api.get_all(as_tuples='type')
+    result = api.get_all(as_values='type')
     assert result == [
-        ('type1', ),
-        ('type2', ),
+        'type1',
+        'type2',
     ]